Refine pool allocator jetton payload tests
