+++ dynamic-capital-ton/apps/tests/pool_allocator.test.ts
@@ -3,33 +3,16 @@ import {
   assertThrows,
 } from "https://deno.land/std@0.224.0/assert/mod.ts";
 import { Buffer } from "node:buffer";
+import { beginCell, toNano } from "npm:@ton/core";
 import {
-  Address,
-  Cell,
-  beginCell,
-  toNano,
-} from "npm:@ton/core";
-
-const OP_JETTON_TRANSFER = 0x0f8a7ea5;
-const OP_DEPOSIT = 0x504f4f4c;
-
-interface DepositForwardPayload {
-  depositId: bigint;
-  investorKey: string;
-  usdtAmount: bigint;
-  dctAmount: bigint;
-  expectedFx: bigint;
-  tonTxHash: string;
-}
-
-interface JettonTransferBody {
-  jettonAmount: bigint;
-  destination: Address;
-  responseDestination: Address;
-  forwardTonAmount: bigint;
-  forwardPayload: Cell;
-  queryId?: bigint;
-}
+  DEFAULT_FORWARD_DESTINATION,
+  DEFAULT_FORWARD_RESPONSE,
+  OP_JETTON_TRANSFER,
+  createDepositForwardPayload,
+  createJettonTransferBody,
+  decodeAllocatorForwardPayload,
+  normalizeHex,
+} from "./helpers/ton.ts";
 
 interface SwapInput {
   depositId: string;
@@ -57,96 +40,6 @@ interface JettonTransferInput extends SwapInput {
   wallet: string;
   jettonAmount: number;
   forwardTonAmount: number;
-}
-
-const DEFAULT_FORWARD_DESTINATION = Address.parse(
-  "0:1111111111111111111111111111111111111111111111111111111111111111",
-);
-const DEFAULT_FORWARD_RESPONSE = Address.parse(
-  "0:2222222222222222222222222222222222222222222222222222222222222222",
-);
-
-function normalizeHex(hex: string, bytes = 32): string {
-  let normalized = hex.startsWith("0x") ? hex.slice(2) : hex;
-  if (normalized.length > bytes * 2) {
-    throw new Error("hex value exceeds expected length");
-  }
-  normalized = normalized.padStart(bytes * 2, "0");
-  return `0x${normalized.toLowerCase()}`;
-}
-
-function hexToBytes(hex: string, bytes = 32): Uint8Array {
-  const normalized = normalizeHex(hex, bytes).slice(2);
-  const result = new Uint8Array(bytes);
-  for (let i = 0; i < bytes; i++) {
-    const start = i * 2;
-    result[i] = parseInt(normalized.slice(start, start + 2), 16);
-  }
-  return result;
-}
-
-function bytesToHex(bytes: Uint8Array): string {
-  return `0x${Array.from(bytes).map((value) => value.toString(16).padStart(2, "0")).join("")}`;
-}
-
-function createDepositForwardPayload(
-  payload: DepositForwardPayload,
-): Cell {
-  const builder = beginCell();
-  builder.storeUint(OP_DEPOSIT, 32);
-  builder.storeUint(payload.depositId, 64);
-  builder.storeBuffer(Buffer.from(hexToBytes(payload.investorKey)));
-  builder.storeCoins(payload.usdtAmount);
-  builder.storeCoins(payload.dctAmount);
-  builder.storeUint(payload.expectedFx, 64);
-  builder.storeBuffer(Buffer.from(hexToBytes(payload.tonTxHash)));
-  return builder.endCell();
-}
-
-function createJettonTransferBody({
-  jettonAmount,
-  destination,
-  responseDestination,
-  forwardTonAmount,
-  forwardPayload,
-  queryId = 0n,
-}: JettonTransferBody): Cell {
-  const builder = beginCell();
-  builder.storeUint(OP_JETTON_TRANSFER, 32);
-  builder.storeUint(queryId, 64);
-  builder.storeCoins(jettonAmount);
-  builder.storeAddress(destination);
-  builder.storeAddress(responseDestination);
-  builder.storeMaybeRef(null);
-  builder.storeCoins(forwardTonAmount);
-  builder.storeRef(forwardPayload);
-  return builder.endCell();
-}
-
-function decodeAllocatorForwardPayload(cell: Cell) {
-  const slice = cell.beginParse();
-  const op = slice.loadUint(32);
-  if (op !== OP_DEPOSIT) {
-    throw new Error("allocator: unsupported op");
-  }
-  const depositId = slice.loadUintBig(64);
-  const investorKey = bytesToHex(slice.loadBuffer(32));
-  const usdtAmount = slice.loadCoins();
-  const dctAmount = slice.loadCoins();
-  if (dctAmount <= 0n) {
-    throw new Error("allocator: invalid dct amount");
-  }
-  const expectedFx = slice.loadUintBig(64);
-  const tonTxHash = bytesToHex(slice.loadBuffer(32));
-  slice.endParse();
-  return {
-    depositId,
-    investorKey,
-    usdtAmount,
-    dctAmount,
-    expectedFx,
-    tonTxHash,
-  };
 }
 
 class MockAllocator {
@@ -490,16 +383,33 @@ Deno.test("forward payload rejects non-deposit opcodes", () => {
   const invalidForward = beginCell()
     .storeUint(0xdeadbeef, 32)
     .storeUint(1n, 64)
-    .storeBuffer(Buffer.from(hexToBytes("0x0")))
+    .storeBuffer(Buffer.from(new Uint8Array(32)))
     .storeCoins(1n)
     .storeCoins(1n)
     .storeUint(1n, 64)
-    .storeBuffer(Buffer.from(hexToBytes("0x1")))
+    .storeBuffer(Buffer.from(new Uint8Array(32).fill(1)))
     .endCell();
 
   assertThrows(
     () => decodeAllocatorForwardPayload(invalidForward),
     Error,
     "allocator: unsupported op",
+  );
+});
+
+Deno.test("forward payload rejects zero dct amount", () => {
+  const forwardPayload = createDepositForwardPayload({
+    depositId: 1n,
+    investorKey: "0xabc",
+    usdtAmount: 10n,
+    dctAmount: 0n,
+    expectedFx: 1n,
+    tonTxHash: "0xdef",
+  });
+
+  assertThrows(
+    () => decodeAllocatorForwardPayload(forwardPayload),
+    Error,
+    "allocator: invalid dct amount",
   );
 });
